move button to BoardEditor
